Module rest_tensors::matrix::matrix_blas_lapack

Functions

_dgemm

an efficient dgemm manipulation for the matrices equipped with the BasicMatrix trait:

_dgemm_nn

_dgemm_nn_serial

_dgemm_tn

_dgemm_tn_serial

_dgemm_tn_v02

_dpotrf

computes the Cholesky factorization of a real symmetric positive definite matrix A

_dsyev

computes all eigenvalues and , optionally, eigenvectors of a real symmetric matrix A

_einsum_01

_einsum_02

contract_vxc_0_serial

An matrix operation for the Vxc preparation: Contraction of a slice with a full matrix (mat_b[num_basis, num_grid], slice_c[num_grid]) -> mat_a[num_basis, num_grid] einsum(ij,i->ij)

general_check_shape