Function rest_tensors::matrix::matrix_blas_lapack::_dpotrf

pub fn _dpotrf<'a, T>(matr_a: &mut T, uplo: char)where
    T: BasicMatrix<'a, f64>,

computes the Cholesky factorization of a real symmetric positive definite matrix A

uplo: char, U or L
= U: Upper triangle of A is stored
= V: Lower triangle of A is stored

Note: _dpotrf destroys the input matrix matr_a, which, on exist is the factor ‘U’ or ‘L’ from the Cholesky factorizatoin A = UTU or A = LLT

Example

   use rest_tensors::{MatrixFull, MatrixUpper};
   use rest_tensors::matrix::matrix_blas_lapack::_dpotrf;
   // generate a matrix with only the upper triangle of A is stored in `MatrixUpper`.
   let matr_0 = MatrixUpper::from_vec(6, vec![4.0,12.0,37.0,-16.0,-43.0,98.0]).unwrap();
   // transfer it to a MatrixFull format and store in matr_a
   let mut matr_a = matr_0.to_matrixfull().unwrap();
   //println!("{:?}", &matr_a);
   //            |  4.0 | 12.0 |-16.0 |
   // matr_a =   | 12.0 | 37.0 |-43.0 |
   //            |-16.0 |-43.0 | 98.0 |
   _dpotrf(&mut matr_a, 'U');
   let U = matr_a.iter_matrixupper().unwrap().map(|x| *x).collect::<Vec<f64>>();
   assert_eq!(U, vec![2.0,6.0,1.0,-8.0,5.0,3.0]);