Trait rest_tensors::tensor_basic_operation::TensorSliceMutUncheck

pub trait TensorSliceMutUncheck<'a, T>where
    Self: TensorIndexUncheck + TensorOptMutUncheck<'a, T>,{
    fn get2d_slice_mut_uncheck(
        &mut self,
        position: [usize; 2],
        length: usize
    ) -> Option<&mut [T]> { ... }
    fn get4d_slice_mut_uncheck(
        &mut self,
        position: [usize; 4],
        length: usize
    ) -> Option<&mut [T]> { ... }
}

Define for upper-formated tensors specifically, like ERIFold4 with the indice of [i,j,k,l], and MatrixUpper with the indice of [i,j]. According to the upper-formated tensor definition, in principle, i=<j and k<=l. WARNING: For efficiency, “uncheck” here means that we don’t check the index order during the tensor operations. WARNING: As a result, a wrong index order could lead to a wrong operation.

Provided Methods

fn get2d_slice_mut_uncheck(
&mut self,
position: [usize; 2],
length: usize
) -> Option<&mut [T]>

fn get4d_slice_mut_uncheck(
&mut self,
position: [usize; 4],
length: usize
) -> Option<&mut [T]>